requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==40.0.2
Django==4.0
django-cors-headers==3.14.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.2
importlib-metadata==6.1.0
Markdown==3.4.1
Pillow==9.5.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
zipp==3.15.0
